Lift the isEditing states up to App
App has a state that keeps track of the CV being
ready to generate or not based on if all fields
are submitted. Passing a handler from App to Form
to update the state caused an error/warning:
"Warning: Cannot update a component (`App`)
while rendering a different component (`Form`)."
Lifting the states up to App fixes this problem.